requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
comtypes==1.4.15
contourpy==1.3.2
cryptography==46.0.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.4.0
google-auth==2.48.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.15.1
idna==3.11
joblib==1.5.3
keras==2.13.1
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.9
ml_dtypes==0.5.4
namex==0.1.0
numpy==1.24.3
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.13.0.90
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pillow==12.1.0
protobuf==3.20.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.99
pywin32==311
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.1
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.10.1
six==1.17.0
sounddevice==0.5.5
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
threadpoolctl==3.6.0
typing_extensions==4.5.0
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==2.0.1
